fix: not-found nitpicks
